requirements.txt
Flask==3.0.0
flask-cors==4.0.0
groq==0.9.0
PyPDF2==3.0.1
Pillow==10.1.0
pytesseract==0.3.10
Werkzeug==3.0.1
